refactored Elevator subsytem to be under subsystem folder not swerve
