+++ src/main/java/frc/robot/commands/ManualElevatorControl.java
@@ -6,8 +6,10 @@ package frc.robot.commands;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.wpilibj2.command.Command;
+import frc.robot.Constants.ElevatorConstants;
 import frc.robot.Constants.OperatorConstants;
-import frc.robot.subsystems.swervedrive.Elevator;
+import frc.robot.subsystems.Elevator;
+
 import java.util.function.DoubleSupplier;
 
 
@@ -30,12 +32,15 @@ public class ManualElevatorControl extends Command {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    elevator.manualMove(MathUtil.applyDeadband(joystick.getAsDouble(),OperatorConstants.LEFT_Y_DEADBAND)* .4);
+    elevator.manualMove(MathUtil.applyDeadband(joystick.getAsDouble(),OperatorConstants.LEFT_Y_DEADBAND)
+                                                                     * ElevatorConstants.JStick_Speed_Mult);
   }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    elevator.stop();
+  }
 
   // Returns true when the command should end.
   @Override
